✨ feat(LotDialog): add new component for lot creation and editing
- Introduce LotDialog component to handle lot creation and editing with form validation
- Implement hierarchical selection for buildings, parts, and part floors
- Enhance user experience with dynamic updates to available options based on selections
- Integrate with existing LotsTab component for seamless lot management
